Move astro.h/astro.cpp to new celastro directory, split functionality
- On Windows, get month abbreviations from the locale instead of the translations
