+++ src/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(CELESTIA_LIBS
   cel3ds
+  celastro
   celcompat
   celengine
   celephem
