+++ src/celestia/qt/qttimetoolbar.cpp
@@ -19,7 +19,7 @@
 #include <QString>
 #include <QTime>
 
-#include <celengine/astro.h>
+#include <celastro/date.h>
 #include <celestia/celestiacore.h>
 #include <celutil/gettext.h>
